21/9/2019

+++ admin/protected/controllers/UsersController.php
@@ -340,7 +340,7 @@ class UsersController extends Controller
         $this->performAjaxValidation($model);
         if(isset($_POST['Users']))
         {
-            if(!$model->getIsNewRecord() && $_POST['Users']['role'] ==5)
+            if(!$model->getIsNewRecord() && $_POST['Users']['role'] ==5 && Yii::app()->user->getState =='Superuser')
             {
                 $usersgroups=GroupUsers::model()->with('group')->findAll('user_id='.$model->id.' and group.course_id='.base64_decode($_GET['c']));
                 if(!empty($usersgroups))
@@ -387,7 +387,7 @@ class UsersController extends Controller
             $model->attributes=$_POST['Users'];
             $model->email=$_POST['Users']['email'];
             $model->updated_date = date('Y-m-d H:i:s');
-            if($_POST['Users']['role'] ==5 ||  $_POST['Users']['role'] ==3)
+            if(($_POST['Users']['role'] ==5 ||  $_POST['Users']['role'] ==3) && Yii::app()->user->getState =='superuser')
             {
                 $facexpl=implode(",",$_POST['Users']['fac_id']);
                 $couexpl=implode(",",$_POST['Users']['course_id'])    ;
@@ -396,29 +396,32 @@ class UsersController extends Controller
             }
 
             if($model->save()) {
-                $UserFaculties = UserFaculties::model()->findAll('user_id='.$model->id);
-                $UserCourses = UserCourses::model()->findAll('user_id='.$model->id);
-                foreach($UserFaculties as $faculties){
-                    $faculties->delete();
-                }
-                foreach($UserCourses as $courses){
-                    $courses->delete();
-                }
-                foreach($_POST['Users']['fac_id'] as $fac_id){
-                    $UserFaculties = new UserFaculties();
-                    $UserFaculties->user_id = $model->id;
-                    $UserFaculties->faculty_id = $fac_id;
-                    $UserFaculties->save();
-                }
-                foreach($_POST['Users']['course_id'] as $course_id){
-                    $UserCourses = new UserCourses();
-                    $UserCourses->user_id = $model->id;
-                    $UserCourses->course_id = $course_id;
-                    $UserCourses->save();
-                }
-                if(isset($_POST['grp']) && !empty($_POST['grp']) && $_POST['Users']['role'] ==5 )
-                {
-                    $usersgroups=GroupUsers::model()->with('group')->findAll('user_id='.$model->id.'and group.course_id='.base64_decode($_GET['c']));
+                if(Yii::app()->user->getState('role') =='Superuser') {
+                    $course_decode=base64_decode($_GET['c']);
+                    $UserFaculties = UserFaculties::model()->findAll('user_id='.$model->id);
+                    $UserCourses = UserCourses::model()->findAll('user_id='.$model->id." and course_id=".$course_decode);
+                    foreach($UserFaculties as $faculties){
+                        $faculties->delete();
+                    }
+                    foreach($UserCourses as $courses){
+                        $courses->delete();
+                    }
+                    foreach($_POST['Users']['fac_id'] as $fac_id){
+                        $UserFaculties = new UserFaculties();
+                        $UserFaculties->user_id = $model->id;
+                        $UserFaculties->faculty_id = $fac_id;
+                        $UserFaculties->save();
+                    }
+                    foreach($_POST['Users']['course_id'] as $course_id){
+                        $UserCourses = new UserCourses();
+                        $UserCourses->user_id = $model->id;
+                        $UserCourses->course_id = $course_id;
+                        $UserCourses->save();
+                    }
+                    if(isset($_POST['grp']) && !empty($_POST['grp']) && $_POST['Users']['role'] ==5 )
+                    {
+                        $usersgroups=GroupUsers::model()->with('group')->findAll('user_id='.$model->id.'and group.course_id='.base64_decode($_GET['c']));
+                    }
                 }
                 Yii::app()->user->setFlash('success','A user has been updated.');
                 //$this->redirect(array('users/cadmin',"c"=>$_GET['c'],"i"=>$_GET['i'],"f"=>$_GET['f']));
